Always render on interactives on all content types
